restructure initial settings to be outside of load
