+++ tif2blender/t2b_nodes/nodeCrosshatch.py
